% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{run Shiny Example}
\usage{
runShiny(shinyappname = "ContDataSumViz_EMVL")
}
\arguments{
\item{shinyappname}{Shiny appplication name, default = ContDataSumViz_EMVL}
}
\description{
Launches Shiny app for ContDataSumViz package.
}
\details{
The Shiny app based on the R package ContDataSumViz is included in
the R package. This function launches that app.

The Shiny app is online at:
https://tetratech-wtr-wne.shinyapps.io/ContDataSumViz
}
\examples{
\dontrun{
# Run Function (default EMVL)
runShiny()

# Shiny, test
runShiny("ContDataSumViz_test")

}
}
